Fix: checking if cluster_registration_token exists on cluster resource creation
